+++ src/compas/datastructures/halfface/halfface.py
@@ -715,7 +715,7 @@ class HalfFace(Datastructure):
             else:
                 yield cell, self.cell_attributes(cell)
 
-    def vertices_where(self, conditions, data=False):
+    def vertices_where(self, conditions, data=False, **kwargs):
         """Get vertices for which a certain condition or set of conditions is true.
 
         Parameters
@@ -726,6 +726,8 @@ class HalfFace(Datastructure):
             values or ranges of attribute values in the form of min/max pairs.
         data : bool, optional
             If True, yield the vertex attributes in addition to the identifiers.
+        **kwargs : dict[str, Any], optional
+            Additional conditions provided as named function arguments.
 
         Yields
         ------
@@ -734,6 +736,9 @@ class HalfFace(Datastructure):
             If `data` is True, the next vertex and its attributes.
 
         """
+        conditions = conditions or {}
+        conditions.update(kwargs)
+
         for key, attr in self.vertices(True):
             is_match = True
 
@@ -811,7 +816,7 @@ class HalfFace(Datastructure):
                 else:
                     yield key
 
-    def edges_where(self, conditions, data=False):
+    def edges_where(self, conditions, data=False, **kwargs):
         """Get edges for which a certain condition or set of conditions is true.
 
         Parameters
@@ -822,6 +827,8 @@ class HalfFace(Datastructure):
             values or ranges of attribute values in the form of min/max pairs.
         data : bool, optional
             If True, yield the edge attributes in addition to the identifiers.
+        **kwargs : dict[str, Any], optional
+            Additional conditions provided as named function arguments.
 
         Yields
         ------
@@ -830,6 +837,9 @@ class HalfFace(Datastructure):
             If `data` is True, the next edge as a (u, v, data) tuple.
 
         """
+        conditions = conditions or {}
+        conditions.update(kwargs)
+
         for key in self.edges():
             is_match = True
 
@@ -891,7 +901,7 @@ class HalfFace(Datastructure):
                 else:
                     yield key
 
-    def faces_where(self, conditions, data=False):
+    def faces_where(self, conditions, data=False, **kwargs):
         """Get faces for which a certain condition or set of conditions is true.
 
         Parameters
@@ -902,6 +912,8 @@ class HalfFace(Datastructure):
             values or ranges of attribute values in the form of min/max pairs.
         data : bool, optional
             If True, yield the face attributes in addition to the identifiers.
+        **kwargs : dict[str, Any], optional
+            Additional conditions provided as named function arguments.
 
         Yields
         ------
@@ -910,6 +922,9 @@ class HalfFace(Datastructure):
             If `data` is True, the next face and its attributes.
 
         """
+        conditions = conditions or {}
+        conditions.update(kwargs)
+
         for fkey in self.faces():
             is_match = True
 
@@ -971,7 +986,7 @@ class HalfFace(Datastructure):
                 else:
                     yield fkey
 
-    def cells_where(self, conditions, data=False):
+    def cells_where(self, conditions, data=False, **kwargs):
         """Get cells for which a certain condition or set of conditions is true.
 
         Parameters
@@ -982,6 +997,8 @@ class HalfFace(Datastructure):
             values or ranges of attribute values in the form of min/max pairs.
         data : bool, optional
             If True, yield the cell attributes in addition to the identifiers.
+        **kwargs : dict[str, Any], optional
+            Additional conditions provided as named function arguments.
 
         Yields
         ------
@@ -990,6 +1007,9 @@ class HalfFace(Datastructure):
             If `data` is True, the next cell and its attributes.
 
         """
+        conditions = conditions or {}
+        conditions.update(kwargs)
+
         for ckey in self.cells():
             is_match = True